baselines/promptSQL++/requirements.txt
google-genai
openai
python-dotenv
tqdm
